Extract cartocss into its own scss file
To take advantage of SCSS features and to reuse color definitions,
cartocss was moved into a scss file and loaded with raw-loader.

This did not work perfectly, due to incompatibilities between
the CSS and CartoCSS specs regarding attribute selectors.

CSS does not allow one to use the following attribute selectors:

  [attr=42]
  [attr<42]

Its set of operators is limited, and no numbers are allowed in the
right hand side of the expression. To work around this, we write
them as:

  [attr="carto(=,42)"]
  [attr="carto(<,42)"]

This is then transformed via a custom loader that uses a regexp to
do the substitution as a final step, after SCSS has done its job.

Not ideal, but I think it's an acceptable trade-off.